# Add script to show optimal build order to make full use of parallel jobs